% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/el_cols.R
\name{el_cols}
\alias{el_cols}
\title{Set EL colors}
\usage{
el_cols(...)
}
\description{
This function allows you to add EL colors to your ggplotgraphics.
NOTE: EL colors include crimson, dark navy, sky blue, orange, lime, charcoal, cool gray, wood, ivory.
}
\examples{
ggplot(mtcars, aes(hp, mpg)) +
geom_point(color = el_cols("dark navy"),
        size = 4, alpha = .8)
}
