% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/el_pal.R
\name{el_pal}
\alias{el_pal}
\title{Return function to interpolate a EL color palette. Currently the only palette is main,
which is crimson and gray}
\usage{
el_pal(palette = "main", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in el_palettes}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments to pass to colorRampPalette()}
}
\description{
Return function to interpolate a EL color palette. Currently the only palette is main,
which is crimson and gray
}
