% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/el_theme.R
\name{el_theme}
\alias{el_theme}
\title{Add EL theme to ggplot chart}
\usage{
el_theme(..., base_size = 12)
}
\description{
This function allows you to add the EL theme to your ggplotgraphics.
NOTE: you may need to run extrafonts::font_import() and extrafonts::loadfonts() to use the Calibri font.
}
\examples{
line <- ggplot(line_df, aes(x = year, y = lifeExp)) +
geom_line(colour = "#007f7f", size = 1) +
geom_hline(yintercept = 0, size = 1, colour="#333333") +
el_theme()
}
